webapp/lib/python/requirements.txt
numpy
Pillow
onnxruntime
